+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# [1.112.0](https://github.com/chris-yyau/busdriver/compare/v1.111.0...v1.112.0) (2026-08-24)
+
+
+### Features
+
+* **hooks:** add bounded fail-open Read size advisory ([#626](https://github.com/chris-yyau/busdriver/issues/626)) ([#756](https://github.com/chris-yyau/busdriver/issues/756)) ([e88f300](https://github.com/chris-yyau/busdriver/commit/e88f3005d7dea83698a620bdfa9bf40028819f6d))
+
+# [1.111.0](https://github.com/chris-yyau/busdriver/compare/v1.110.14...v1.111.0) (2026-08-24)
+
+
+### Features
+
+* **dispatch-cli:** land pi-read rename and delete the shipped model default ([#755](https://github.com/chris-yyau/busdriver/issues/755)) ([2ca9675](https://github.com/chris-yyau/busdriver/commit/2ca967545c90ec1c775e618c2a8e0314829ba59f))
+
 ## [1.110.14](https://github.com/chris-yyau/busdriver/compare/v1.110.13...v1.110.14) (2026-08-23)